Event Type
Modified Events and Event Dispatcher to allow for enum event types.
This is basically a rewrite, I tried to keep it as true to the original blog post.
My changes could be outlined from https://github.com/Pseudomanifold/Events/issues/1.
These changes made Event and Event Dispatcher into templated classes. These templated classes will accept a `enum` as it's type.
This move the event type to the compiler for validation.